# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PRD Engine API
  description: |
    API for the PRD (Product Requirements Document) Engine. It supports:
    - **Modules**: Create and update documentation modules with versioning. Each save creates a new version; previous versions are kept for history.
    - **Audit**: Every module tracks who created/updated it and when (CreatedBy, UpdatedBy, CreatedAt, UpdatedAt).
    - **Headers**: Extract markdown headings from a module for table-of-contents or in-page navigation.
    - **Export**: Combine all modules (latest versions only) into one document and export as JSON or PDF, including a document history table (who updated what).
    All API routes under `/api` require Bearer token authentication (see Security).
  version: 1.0.0
  contact:
    name: PRD Engine
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /api/modules:
    get:
      tags: [Modules]
      summary: List all modules (latest versions)
      description: Returns a list of all modules with their latest version metadata. Useful for building a sidebar or module index.
      operationId: listModules
      responses:
        '200':
          description: List of modules
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Modules listed successfully
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Module'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: Failed to fetch modules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
    post:
      tags: [Modules]
      summary: Create or update a module (new version)
      description: |
        Creates a new module (version 1) if the module ID does not exist, or saves a **new version** of an existing module (version incremented).
        The request body is the full module payload; the authenticated user (from Bearer token) is recorded as UpdatedBy (and CreatedBy for new modules).
        Order must be a positive integer (used for sorting in combined export).
      operationId: saveModule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveModuleRequest'
      responses:
        '200':
          description: Module saved successfully (new version created)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponseWithModule'
        '400':
          description: Invalid request payload (missing or invalid fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '401':
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '422':
          description: Validation or business rule error (e.g. invalid order)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Token from your auth provider (file-based tokens in config/users.yaml). Send as Authorization: Bearer <token>."
  responses:
    Unauthorized:
      description: Missing or invalid Authorization header
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
  schemas:
    SaveModuleRequest:
      type: object
      required: [id, title, order, surfaces, content]
      properties:
        id:
          type: string
          description: Unique module identifier (e.g. checkout-flow)
        title:
          type: string
          description: Display title
        order:
          type: integer
          minimum: 1
          description: Sort order for combined export (lower = earlier in document)
        surfaces:
          type: array
          items:
            type: string
            enum: [customer_app, website, admin_dashboard, partner_app, leads_app]
          description: Product surfaces this module applies to
        content:
          type: string
          description: Markdown body (can include # ## ### headings)
    Module:
      type: object
      properties:
        ID:
          type: string
        Version:
          type: integer
        Title:
          type: string
        Order:
          type: integer
        Surfaces:
          type: array
          items:
            type: string
        Content:
          type: string
        CreatedBy:
          type: string
        CreatedAt:
          type: string
          format: date-time
        UpdatedBy:
          type: string
        UpdatedAt:
          type: string
          format: date-time
...
    APIResponseWithModule:
      type: object
      properties:
        status:
          type: string
        statusCode:
          type: integer
          example: 200
        message:
          type: string
        data:
          $ref: '#/components/schemas/Module'
...
    APIError:
      type: object
      properties:
        status:
          type: string
        statusCode:
          type: integer
        message:
          type: string
        userMessageTitle:
          type: string
        userMessageText:
          type: string
        errors:
          type: string
          description: Optional validation details
        tag:
          type: string